Similar structure for tests +minor
