+++ tests/Plugins/GitHubModule/RulesetRequirements/RequireConversationResolutionRule_UnitTest.py
@@ -1,0 +1,73 @@
+# -------------------------------------------------------------------------------
+# |
+# |  Copyright (c) 2024 Scientific Software Engineering Center at Georgia Tech
+# |  Distributed under the MIT License.
+# |
+# -------------------------------------------------------------------------------
+"""Unit tests for RequireConversationResolutionRule.py"""
+
+from RepoAuditor.Plugins.GitHub.RulesetRequirements.RequireConversationResolution import (
+    RequireConversationResolutionRule,
+)
+from RepoAuditor.Requirement import EvaluateResult
+
+from . import create_rule
+
+
+class TestRequireConversationResolutionRule:
+    """Unit tests for the RequireConversationResolutionRule requirement."""
+
+    def test_Enabled(self, query_data):
+        """Test if the rule is enabled on the main branch."""
+        requirement = RequireConversationResolutionRule()
+
+        query_data["rules"].append(
+            create_rule(
+                name="PR Rules",
+                rule_type="pull_request",
+                ruleset_name="main",
+                parameters={"required_review_thread_resolution": True},
+            )
+        )
+
+        result = requirement.Evaluate(query_data, {"no": False})
+        assert result.result == EvaluateResult.Success
+        assert f"{requirement.github_settings_value} is enabled" in result.context
+
+    def test_Disabled(self, query_data):
+        """Test if the required rule is disabled on the main branch
+        by providing the value as false.
+        """
+        requirement = RequireConversationResolutionRule()
+
+        query_data["rules"].append(
+            create_rule(
+                name="PR Rules",
+                # Wrong rule type
+                rule_type="pull_request",
+                ruleset_name="main",
+                parameters={"required_review_thread_resolution": False},
+            )
+        )
+
+        result = requirement.Evaluate(query_data, {"no": False})
+        assert result.result == EvaluateResult.Error
+
+    def test_WrongRuleType(self, query_data):
+        """Test if the required rule is disabled on the main branch
+        by providing the wrong rule type.
+        """
+        requirement = RequireConversationResolutionRule()
+
+        query_data["rules"].append(
+            create_rule(
+                name="Inactive PR Rules",
+                # Wrong rule type
+                rule_type="push_request",
+                ruleset_name="Ruleset Main",
+                parameters={},
+            )
+        )
+
+        result = requirement.Evaluate(query_data, {"no": False})
+        assert result.result == EvaluateResult.Error
