+++ tests/Plugins/GitHubModule/StandardRequirements/SquashMergeCommitMessage_UnitTest.py
